+++ lib/danger/request_sources/github.rb
@@ -10,6 +10,14 @@ module Danger
 
       attr_accessor :pr_json, :issue_json, :support_tokenless_auth
 
+      def self.env_vars
+        ["DANGER_GITHUB_API_TOKEN"]
+      end
+
+      def self.optional_env_vars
+        ["DANGER_GITHUB_HOST", "DANGER_GITHUB_API_BASE_URL"]
+      end
+
       def initialize(ci_source, environment)
         self.ci_source = ci_source
         self.environment = environment