+++ spec/lib/danger/request_sources/request_source_spec.rb
@@ -31,4 +31,49 @@ describe Danger::RequestSources::RequestSource, host: :github do
       expect(g.validates_as_ci?).to be false
     end
   end
+
+  describe ".source_name" do
+    context "GitHub" do
+      it "returns the name of request source" do
+        result = Danger::RequestSources::GitHub.source_name
+
+        expect(result).to eq "GitHub"
+      end
+    end
+
+    context "GitLab" do
+      it "returns the name of request source" do
+        result = Danger::RequestSources::GitLab.source_name
+
+        expect(result).to eq "GitLab"
+      end
+    end
+
+    context "BitbucketCloud" do
+      it "returns the name of request source" do
+        result = Danger::RequestSources::BitbucketCloud.source_name
+
+        expect(result).to eq "BitbucketCloud"
+      end
+    end
+
+    context "BitbucketCloud" do
+      it "returns the name of request source" do
+        result = Danger::RequestSources::BitbucketServer.source_name
+
+        expect(result).to eq "BitbucketServer"
+      end
+    end
+  end
+
+  describe ".available_source_names_and_envs" do
+    it "returns list of items contains source name and envs" do
+      result = described_class.available_source_names_and_envs.join(", ")
+
+      expect(result).to include("- GitHub:")
+      expect(result).to include("- GitLab:")
+      expect(result).to include("- BitbucketCloud:")
+      expect(result).to include("- BitbucketServer:")
+    end
+  end
 end